# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "분리수ZIP 서비스 API",
    "description": "엘리스 AI 트랙 4기 - AI Project 10팀 TOPTEN 프로젝트 문서.",
    "version": "1.0.0"
  },
  "host": "home.handwoong.com:5001",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "in": "header",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
...
    "/trash": {
      "get": {
        "tags": [
          "trash"
        ],
        "description": "쓰레기 목록 조회",
        "parameters": [
          {
            "name": "queryString",
            "in": "query",
            "description": "**search** 검색어\n  **category** 조회할 카테고리\n  **page** 첫 요청시 빈 문자열 또는 생략\n  **limit** 기본값10\n",
            "required": false,
            "schema": {
              "$ref": "#/definitions/TrashGetQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/TrashGetResponse"
            },
            "description": "쓰레기 목록을 배열형태로 반환"
          }
        }
      },
      "post": {
        "tags": [
          "trash"
        ],
        "description": "쓰레기 생성",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "생성하고자 하는 쓰레기의 정보를 body에 담아 요청",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TrashPostRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/TrashPostResponse"
            },
            "description": "생성된 쓰레기 정보 반환"
          }
        }
      }
    },
...
    "/trash/{id}": {
      "get": {
        "tags": [
          "trash"
        ],
        "description": "개별 쓰레기 정보 조회",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "얻고자 하는 쓰레기정보의 ID",
            "schema": {
              "$ref": "#/definitions/TrashId"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/TrashOneGetResponse"
            },
            "description": "쓰레기 정보를 반환"
          }
        }
      },
      "put": {
        "tags": [
          "trash"
        ],
        "description": "쓰레기 정보 수정",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "수정하고자 하는 쓰레기의 ID",
            "schema": {
              "$ref": "#/definitions/TrashId"
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "수정하고자 하는 쓰레기의 정보를 body에 담아 요청",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TrashPutRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/TrashPutResponse"
            },
            "description": "수정된 쓰레기 정보 반환"
          }
        }
      },
      "delete": {
        "tags": [
          "trash"
        ],
        "description": "쓰레기 삭제",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "삭제하고자 하는 쓰레기의 ID",
            "schema": {
              "$ref": "#/definitions/TrashId"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/DeleteResponse"
            },
            "description": "삭제 메시지"
          }
        }
      }
    },
    "/trash/ai": {
      "post": {
        "tags": [
          "trash"
        ],
        "description": "쓰레기 AI 분석",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "multFiles",
            "in": "formData",
            "type": "single",
            "required": true,
            "description": "쓰레기 이미지 파일\nmultipart/form-data",
            "collectionFormat": "multi",
            "items": {
              "type": "file"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/TrashAiResponse"
            },
            "description": "AI 분석결과를 반환"
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "한명의 유저 조회",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "얻고자 하는 유저정보의 ID",
            "schema": {
              "$ref": "#/definitions/UserId"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/UserGetResponse"
            },
            "description": "한명의 유저 정보를 반환"
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "유저 생성",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "생성하고자 하는 유저의 정보를 body에 담아 요청\n  **email** 이메일 **필수**\n  **password** 비밀번호 **필수** 8자리 이상\n  **username** 닉네임 **필수** 3자리 이상  ",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/UserGetResponse"
            },
            "description": "생성된 유저 정보 반환"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "유저 로그인",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "유저 로그인 정보를 담아 요청\n  **email** 이메일 **필수**\n  **password** 비밀번호 **필수** 8자리 이상\n  ",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserLoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/UserGetResponse"
            },
            "description": "로그인 유저 정보 반환"
          }
        }
      }
    },
    "/users/update": {
      "put": {
        "tags": [
          "user"
        ],
        "description": "유저 수정",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "수정하고자 하는 유저의 정보를 body에 담아 요청\n  **로그인 필수**\n  **email** 이메일 **필수** 변경 불가\n  **password** 비밀번호 **선택**\n  **username** 닉네임 **선택**  ",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/UserGetResponse"
            },
            "description": "수정된 유저 정보 반환"
          }
        }
      }
    },
    "/users/score": {
      "put": {
        "tags": [
          "user"
        ],
        "description": "유저의 미니게임 점수 갱신",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "수정하고자 하는 유저의 정보를 body에 담아 요청\n  **로그인 필수**\n  **score** 점수 **필수**  ",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserScoreRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/UserScoreResponse"
            },
            "description": "점수 갱신 메시지 반환"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DeleteResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "삭제가 완료되었습니다."
        }
      }
    },
    "UserId": {
      "type": "string",
      "example": "users/{userId}"
    },
    "UserGetResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "62acd51598886269f298521e"
        },
        "email": {
          "type": "string",
          "example": "test@test.com"
        },
        "username": {
          "type": "string",
          "example": "테스트"
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      }
    },
    "UserRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "test@test.com"
        },
        "password": {
          "type": "string",
          "example": "test1234"
        },
        "username": {
          "type": "string",
          "example": "테스트유저"
        }
      }
    },
    "UserLoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "test@test.com"
        },
        "password": {
          "type": "string",
          "example": "test1234"
        }
      }
    },
...
    "UserScoreRequest": {
      "type": "object",
      "properties": {
        "score": {
          "type": "number",
          "example": 80
        }
      }
    },
    "UserScoreResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "점수 갱신이 완료되었습니다."
        }
      }
    },
...
    "TrashId": {
      "type": "string",
      "example": "trash/{trashId}"
    },
    "TrashGetQuery": {
      "type": "string",
      "example": "?search={}&category={}&page={ID}&limit={5}"
    },
    "TrashGetResponse": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "62a1624d1458dc8c48ab52ca"
          },
          "title": {
            "type": "string",
            "example": "컵라면"
          },
          "description": {
            "type": "object",
            "properties": {
              "throwAway": {
                "type": "array",
                "example": [
                  "컵라면 용기를 물로 세척",
                  "햇빛에 하루 이상 말려 기름, 국물 자국 제거",
                  "흰 색의 용기로 만든 뒤 배출"
                ],
                "items": {
                  "type": "string"
                }
              },
              "note": {
                "type": "array",
                "example": [
                  "라면 국물이 남아있거나, 음식물이 완전히 제거 되지 않은 컵라면 용기는 재활용 불가능하므로 일반쓰레기로 배출"
                ],
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "kind": {
            "type": "array",
            "example": [
              "다양한 스티로폼 컵라면 용기"
            ],
            "items": {
              "type": "string"
            }
          },
          "image": {
            "type": "string",
            "example": "https://upload.wikimedia.org/wikipedia/commons/0/0a/PSP%EB%9D%BC%EB%A9%B4%EC%9A%A9%EA%B8%B0.png"
          },
          "recycle": {
            "type": "boolean",
            "example": true
          },
          "category": {
            "type": "array",
            "example": [
              "스티로폼"
            ],
            "items": {
              "type": "string"
            }
          },
          "__v": {
            "type": "number",
            "example": 0
          }
        }
      }
    },
    "TrashAiResponse": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "페트병"
        },
        "kind": {
          "type": "string",
          "example": "플라스틱"
        },
        "section": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "example": "라벨"
              },
              "score": {
                "type": "number",
                "example": 0
              }
            }
          }
        },
        "throwAway": {
          "type": "array",
          "example": [
            "내용물을 비운 뒤 세척",
            "부착 상표 등을 제거 후 일반 쓰레기에 버리기"
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "TrashOneGetResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "62a1624d1458dc8c48ab52ca"
        },
        "title": {
          "type": "string",
          "example": "컵라면"
        },
        "description": {
          "type": "object",
          "properties": {
            "throwAway": {
              "type": "array",
              "example": [
                "컵라면 용기를 물로 세척",
                "햇빛에 하루 이상 말려 기름, 국물 자국 제거",
                "흰 색의 용기로 만든 뒤 배출"
              ],
              "items": {
                "type": "string"
              }
            },
            "note": {
              "type": "array",
              "example": [
                "라면 국물이 남아있거나, 음식물이 완전히 제거 되지 않은 컵라면 용기는 재활용 불가능하므로 일반쓰레기로 배출"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        },
        "kind": {
          "type": "array",
          "example": [
            "다양한 스티로폼 컵라면 용기"
          ],
          "items": {
            "type": "string"
          }
        },
        "image": {
          "type": "string",
          "example": "https://upload.wikimedia.org/wikipedia/commons/0/0a/PSP%EB%9D%BC%EB%A9%B4%EC%9A%A9%EA%B8%B0.png"
        },
        "recycle": {
          "type": "boolean",
          "example": true
        },
        "category": {
          "type": "array",
          "example": [
            "스티로폼"
          ],
          "items": {
            "type": "string"
          }
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      }
    },
...
    "TrashPostRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "컵라면"
        },
        "description": {
          "type": "object",
          "properties": {
            "throwAway": {
              "type": "array",
              "example": [
                "컵라면 용기를 물로 세척",
                "햇빛에 하루 이상 말려 기름, 국물 자국 제거",
                "흰 색의 용기로 만든 뒤 배출"
              ],
              "items": {
                "type": "string"
              }
            },
            "note": {
              "type": "array",
              "example": [
                "라면 국물이 남아있거나, 음식물이 완전히 제거 되지 않은 컵라면 용기는 재활용 불가능하므로 일반쓰레기로 배출"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        },
        "kind": {
          "type": "array",
          "example": [
            "다양한 스티로폼 컵라면 용기"
          ],
          "items": {
            "type": "string"
          }
        },
        "image": {
          "type": "string",
          "example": "https://upload.wikimedia.org/wikipedia/commons/0/0a/PSP%EB%9D%BC%EB%A9%B4%EC%9A%A9%EA%B8%B0.png"
        },
        "recycle": {
          "type": "boolean",
          "example": true
        },
        "category": {
          "type": "array",
          "example": [
            "스티로폼"
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "TrashPostResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "62a1624d1458dc8c48ab52ca"
        },
        "title": {
          "type": "string",
          "example": "컵라면"
        },
        "description": {
          "type": "object",
          "properties": {
            "throwAway": {
              "type": "array",
              "example": [
                "컵라면 용기를 물로 세척",
                "햇빛에 하루 이상 말려 기름, 국물 자국 제거",
                "흰 색의 용기로 만든 뒤 배출"
              ],
              "items": {
                "type": "string"
              }
            },
            "note": {
              "type": "array",
              "example": [
                "라면 국물이 남아있거나, 음식물이 완전히 제거 되지 않은 컵라면 용기는 재활용 불가능하므로 일반쓰레기로 배출"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        },
        "kind": {
          "type": "array",
          "example": [
            "다양한 스티로폼 컵라면 용기"
          ],
          "items": {
            "type": "string"
          }
        },
        "image": {
          "type": "string",
          "example": "https://upload.wikimedia.org/wikipedia/commons/0/0a/PSP%EB%9D%BC%EB%A9%B4%EC%9A%A9%EA%B8%B0.png"
        },
        "recycle": {
          "type": "boolean",
          "example": true
        },
        "category": {
          "type": "array",
          "example": [
            "스티로폼"
          ],
          "items": {
            "type": "string"
          }
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      }
    },
    "TrashPutRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "수정된제목"
        },
        "description": {
          "type": "object",
          "properties": {
            "throwAway": {
              "type": "array",
              "example": [
                "수정",
                "햇빛에 하루 이상 말려 기름, 국물 자국 제거",
                "흰 색의 용기로 만든 뒤 배출"
              ],
              "items": {
                "type": "string"
              }
            },
            "note": {
              "type": "array",
              "example": [
                "라면 국물이 남아있거나, 음식물이 완전히 제거 되지 않은 컵라면 용기는 재활용 불가능하므로 일반쓰레기로 배출"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        },
        "kind": {
          "type": "array",
          "example": [
            "다양한 스티로폼 컵라면 용기"
          ],
          "items": {
            "type": "string"
          }
        },
        "image": {
          "type": "string",
          "example": "https://upload.wikimedia.org/wikipedia/commons/0/0a/PSP%EB%9D%BC%EB%A9%B4%EC%9A%A9%EA%B8%B0.png"
        },
        "recycle": {
          "type": "boolean",
          "example": true
        },
        "category": {
          "type": "array",
          "example": [
            "스티로폼"
          ],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "TrashPutResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "62a1624d1458dc8c48ab52ca"
        },
        "title": {
          "type": "string",
          "example": "수정된제목"
        },
        "description": {
          "type": "object",
          "properties": {
            "throwAway": {
              "type": "array",
              "example": [
                "수정",
                "햇빛에 하루 이상 말려 기름, 국물 자국 제거",
                "흰 색의 용기로 만든 뒤 배출"
              ],
              "items": {
                "type": "string"
              }
            },
            "note": {
              "type": "array",
              "example": [
                "라면 국물이 남아있거나, 음식물이 완전히 제거 되지 않은 컵라면 용기는 재활용 불가능하므로 일반쓰레기로 배출"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        },
        "kind": {
          "type": "array",
          "example": [
            "다양한 스티로폼 컵라면 용기"
          ],
          "items": {
            "type": "string"
          }
        },
        "image": {
          "type": "string",
          "example": "https://upload.wikimedia.org/wikipedia/commons/0/0a/PSP%EB%9D%BC%EB%A9%B4%EC%9A%A9%EA%B8%B0.png"
        },
        "recycle": {
          "type": "boolean",
          "example": true
        },
        "category": {
          "type": "array",
          "example": [
            "스티로폼"
          ],
          "items": {
            "type": "string"
          }
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      }
    },
...
  }
}